+++ fidget/src/jit/x86_64/mod.rs
@@ -12,7 +12,7 @@
 //! available.
 
 /// We use `xmm4-16` (all caller-saved) for graph variables
-pub const REGISTER_LIMIT: u8 = 12;
+pub const REGISTER_LIMIT: usize = 12;
 /// `xmm0` is used for immediates
 pub const IMM_REG: u8 = 0;
 /// `xmm1-3` are available for use as temporaries.
